IMPALA-4880: Clarify synchronization policy for 'done_' in KuduScanNode
This patch adds a comment to clarify that it's safe to optimistically
read 'done_' in some places in KuduScanNode. This is to avoid some
confusion as the code synchronizes access to the variable in certain
places but not in others.

Also, did some miscellaneous cleanup in a couple of places.

Change-Id: I72c5f1fecf4bcf737d71a1fa13e59361604485f0
Reviewed-on: http://gerrit.cloudera.org:8080/5494
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Impala Public Jenkins