IMPALA-4676: remove vestigial references to getBlockStorageLocations() API
The BlockStorageLocation import is unused.

Remove validation of config keys that only affect the
BlockStorageLocation API. See HDFS-10868 and HDFS-8895. We do not
need to validate these keys any more since we don't use that API.
These config keys are removed in Hadoop 3 so this patch is
required to build against it.

Change-Id: Ic12337a9f5b7d910282aaf7d8508a4176cf89cbc
Reviewed-on: http://gerrit.cloudera.org:8080/5526
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins