IMPALA-3401: Suppress blocks of content devoted to Cloudera Manager.
Initial experiment to see how the doc looks & builds when the biggest,
most obvious pieces of Cloudera Manager-related content are suppressed:
- Primarily entire subtopics.
- In some cases, individual notes, paragraphs, and list items.
- Also the CM reference in the reused "cancel" text which showed up
  many times throughout the SQL syntax section.

Just hiding in the output for now, rather than removing from source,
because we might find there's generic material we need to hoist out of
the CM-related subtopics.

Was planning to use audience="hidden" to hide, but since that CR hasn't
landed in master yet, still using audience="Cloudera" for the moment,
and will switch attribute values later.
A little rewording of titles and suppressing smaller elements such as
paragraphs where the CM aspect didn't apply to the entire subtopic.

Change-Id: Ic799f77bc758a1cf40c53fa412bf02d852901a69
Reviewed-on: http://gerrit.cloudera.org:8080/5607
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: John Russell <jrussell@cloudera.com>