Make sure 'errata' page is hidden.
The errata page is in an early experimental form and not expected
to be ready for prime time anytime soon. Depends on some
automation that's not practical to put in place during the
transition to Apache infra. That page is chock full of references
to CDH version numbers, so best to make sure it can't possibly
appear in the output.

Using 'Cloudera' condition instead of 'pre-release' to clarify that
the errata topic is experimental and will remain invisible for the
foreseeable future.

Also applying the same attribute to the corresponding <keydef> tag
so the hidden topic doesn't get processed in some unforeseen way.

Will change those conditions to 'hidden' after the other changes
to audience= attributes land.

Change-Id: Id254192b325ba18c415a2cf17e6357800d81c437
Reviewed-on: http://gerrit.cloudera.org:8080/5625
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Reviewed-by: Ambreen Kazi <ambreen.kazi@cloudera.com>
Tested-by: John Russell <jrussell@cloudera.com>