Initial commit of the blog section of the Impala ASF website.
Note: a sample version of the site is up and running at:

   http://dknupp.github.io/blog/

The blog site is generated using Nikola. Nikola is only being used
to manage the blog in this patch, but could easily be used to manage
the other pages as well. It may take a fair bit of tweaking to the
settings and Jinja templates to make everything work properly if Nikola
is to be used for non-blog pages -- a job for when more cycles are
generally available.

The README gives a pretty complete description of the workflow for
adding content to the blog. The formatted README file can be seen at:

https://github.com/dknupp/dknupp.github.io/tree/master/nikola_site_generator

Comments on blog posts is being handled via Disqus. Other than that,
not much effort was spent in setting up some of the typical bloggish
bells and whistles - e.g. RSS/ATOM feeds, etc. Fancier bits can come
later.

Change-Id: If5810f04b56c587a5114953cae7e4f484a92be75
Reviewed-on: http://gerrit.cloudera.org:8080/5667
Reviewed-by: Jim Apple <jbapple-impala@apache.org>
Tested-by: Jim Apple <jbapple-impala@apache.org>